+++ ex03/ft_strncat.c
@@ -6,21 +6,21 @@
 /*   By: lbiasuz <lbiasuz@student.42sp.org.br>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 20:47:18 by lbiasuz           #+#    #+#             */
-/*   Updated: 2022/02/14 20:51:51 by lbiasuz          ###   ########.fr       */
+/*   Updated: 2022/02/15 15:01:04 by lbiasuz          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char *ft_strncat(char *dest, char *src, unsigned int nb)
 {
-	int	i;
-	int	j;
+	unsigned int	i;
+	unsigned int	j;
 	
 	j = 0;
 	i = 0;
 	while (dest[i] != '\0')
 		i++;
 	dest[i]++;
-	while (src[j] != '\0' && j < (int) nb)
+	while (src[j] != '\0' && j < nb)
 	{
 		dest[i] = src[j];
 		j++;
